Returns dict result to tab by re-structuring and introducing jQuery.
jQuery is used to parse the HTML output from Yahoo dictionary.
I think I will use it to show the preview tooltip as well.

background.js and content.js are also re-designed.
content.js simply informs background.js what's been done in this tab.
It no longer asks background.js to "newtab".  Instead, it indicates it's a
selection event, and what's been selected, which key has been holded during selection,
what are the current tab url, etc.

background.js handles most of the logic, including determining which BE to audit,
which action to do (new tab or show preview), and telling content.js what to show in preview.

This commit allows users to mouse-select (still holding 'c')
and it will hit Yahoo dictionary, parse the title of the first word,
and return to content.js some information.  The rendering of tooltip
and the parsing of explanation of the word are not done in this commit.